Replace custom BVH with tinybvh-rs and optimize traversal
- Vendored `tinybvh-rs` to `crates/tinybvh` and patched `build.rs` to fix cross-compilation (removed `-march=native` and used `is_ok_and`).
- Fixed clippy warnings in `crates/tinybvh/build.rs`, `crates/tinybvh/src/cxx_ffi.rs` and `crates/tinybvh/ffi/tinybvh/tiny_bvh.h`.
- Updated `crates/bvh/Cargo.toml` to depend on the local `tinybvh` crate.
- Added `crates/tinybvh` to workspace members.
- Updated `BVHTree::new` to build BVH from AABBs using 2 triangles per AABB and `split_leaves(1)` to ensure correct 1-to-1 mapping.
- Updated `GPUBVHNode` structure to use `left_node` (index) and `primitive_index` instead of `miss` links.
- Updated `create_linearized_bvh` to linearize `tinybvh` nodes.
- Updated `GlobalBuffers` to correctly offset `left_node` when merging BVHs.
- Updated `common.inc` and `raytracing.inc` shaders to use the new `BVHNode` structure and stack-based traversal (Left/Right child) instead of roped traversal (Miss link).

+++ crates/tinybvh/src/lib.rs
@@ -24,8 +24,6 @@ mod layouts;
 mod ray;
 mod traversal;
 
-use std::u32;
-
 pub(crate) use cxx_ffi::ffi;
 pub use layouts::*;
 pub use ray::*;
@@ -59,7 +57,7 @@ pub enum Error {
 impl Error {
     pub(crate) fn validate_triangulated(prim_len: usize) -> Result<(), Error> {
         if prim_len % 3 != 0 {
-            Err(Error::PrimitiveTriangulated(prim_len as usize))
+            Err(Error::PrimitiveTriangulated(prim_len))
         } else {
             Ok(())
         }
